drm/amdgpu/UAPI: add SENSOR INFO ioctl subquery for profile modes
SQTT (Shader sQuencer Thread Trace) is a hardware feature
that dumps fine grained thread data from the shader hardware
which can then be analyzed using tools like Radeon GPU Profiler
(RGP).  For the thread trace to be useful, we need to make sure
the GPU is set to a profiling mode with consistent clocks and
things like gfxoff and clock and power gating disabled. Add
a query to check if we are in a profiling mode so applications
like mesa can query this before enabling SQTT or SPM (Streaming
Performance Monitor)

Proposed mesa changes:
XXX: todo

Signed-off-by: Alex Deucher <alexander.deucher@amd.com>