libata: add horkage for missing Identify Device log
ACS defines the ATA Identify Device Data log as mandatory. A warning
message currently signals to the user if a device does not report
supporting this log page in the log directory page. However, it is
useless to constantly access the log directory and warn about this lack
of support once we have discovered that the device does not support
this mandatory log page.

Introduce the horkage flag ATA_HORKAGE_NO_ID_DEV_LOG to mark a device as
lacking support for the Identify Device Data log page. Set this flag
when ata_log_supported() returns false in ata_identify_page_supported().
The warning is printed only once on device scan and the log directory
not accessed again to test for Identify Device Data log page support.

Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>