gfs2: Prevent endless loops in gfs2_file_buffered_write
Currently, instead of performing a short write,
iomap_file_buffered_write will fail when part of the iov iterator it is
expected to read from cannot be read.  On the other hand,
gfs2_file_buffered_write will loop around if it can read part of the iov
iterator, so we can end up in an endless loop.

This should be fixed in iomap_file_buffered_write (and also
generic_perform_write) but it's a bit late for that in the 5.16
development cycle, so work around it in the filesystem for now.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>